Ailment_Report_Assistant/requirements.txt
streamlit==1.45.0
google-cloud-bigquery==3.21.0
google-generativeai==0.3.2
pandas==2.2.1
numpy==1.26.0
python-dotenv==1.0.1
pyarrow==14.0.1
db-dtypes==1.1.1
google-auth==2.23.0
google-auth-oauthlib==1.1.0

protobuf==4.25.3
requests==2.32.0
tqdm==4.66.3
cachetools==5.3.1